Add TSLint to the workflow
Configure TSLint checking for common problems.

There is only a single task to lint all source files in one go.
This is due to the limitation of tslint formatter which is not able
to always produce absolute paths, and the limitation of VSCode+lerna
which cannot detect the individual package directory where a
per-package linter would be run via `npm run lint`. As a result, a path
relative to package root (e.g. "packages/loopback") cannot be converted
to an absolute path to offer in the Problems window.

Usage options:

 - `npm run lint` at repo root
 - `npm run test` runs linter after the tests too
 - Run "Lint all files" task in VS code. Check the Problems window
   for the list of issues to fix
 - Install "vscode-tslint" to your VisualStudio Code to get real-time
   checking of a subset of linting rules

As part of this patch, I fixed missing type declarations in
"bootstrapping.ts" test and also fixed "client.ts" to resolve the
returned promise to a value which honours the declared interface
("status" vs. "statusCode" property).