Remove promise-function-async
I enabled promise-function-async to catch the kind of errors
where a function returning a promise is not annotated with type
information and therefore the compiler cannot verify the correctness
of what the function is actually returning.

With noImplicitAny enabled, we have a better tool to detect such errors.

I also want to avoid the following downside of promise-function-async:
it forces async functions even in places where an async function is
clearly a wrong thing to do, e.g.

    const bar = () => Promise.resolve(1);