context: `add @inject.setter`
Implement a new decorator for injecting a setter function
for (re)binding a new value for the given key.

This is useful when implementing Actions that are contributing
new Elements. Compared to existing `bindElement`, this new setter
approach leads to a less coupled code, because the action
invoking the setter does not need to know about the key associated
with the binding, in fact a setter function can be used without any
Context/Binding classes. This also makes Actions easier to test in
isolation.