build: skip CI of feature branches with an open PR
For pull requests sent from a feature branch living in the same
repository, both AppVeyor and Travis CI is running two CI jobs,
which is unnecessarily slowing down our feedback cycle:

 - continuous-integration/appveyor/branch
 - continuous-integration/appveyor/pr

and

 - continuous-integration/travis-ci/pr
 - continuous-integration/travis-ci/push

This commit changes AppVeyor config to skip CI for the branch
and run it for the pull request only.

In order to achieve the same on Travis, we are restricting the branches
built by CI to "master" only. (This means we will need to update Travis
config when forking a long-living "N.x" branch.)