docs: improve PULL_REQUEST_TEMPLATE
Remove the initial boilerplate that very few people are using and that
does not work at all with the default GitHub workflow where the
description of a new pull request is pre-filled from the commit message.

Add more items to the Checklist section to remind both the contributors
and the reviewers to check for commonly forgotten things like API docs.