fix: apply source-maps to test errors
When we got rid of ts-node in #602, we also unintentionally dropped
the code modifying error stack traces to point to original TypeScript
sources instead of the transpiled JavaScript output.

As a result, assertion failures contained stack traces pointing to code
that we are not familiar with, which made troubleshooing difficult.

This commits add source-map-support as a global dev-dependency
plus a dev-dependency of example repositories that must work outside
of our monorepo too.

The single shared `mocha.opts` was split into two different options
files:

 - `mocha.ts.opts` which registers source-map support
 - `mocha.js.opts` with no support for source maps