ci: fix appveyor
AppVeyor builds are intermittenly failing. It seems to me that the issue is manifesting intermittenly when we do `npm install` (which works fine) but then fails when we run `npm run bootstrap` which tries to install stuff again.

Reference threads (not the exact same issue but gives enough hints for me to suspect the double install being the culprit):
https://github.com/okonet/lint-staged/issues/274
https://github.com/npm/npm/issues/18380