feat(context): typed binding keys
Modify Context, Binding, and related methods to allow developers
to add a type information to individual binding keys. These binding
keys will allow the compiler to verify that a correct value is passed
to binding setters (`.to()`, `.toClass()`, etc.) and automatically
infer the type of the value returned by `ctx.get()` and `ctx.getSync()`.

Bring the documentation in sync with the actual implementation in code.
Rename `AuthenticationProvider` to `AuthenticateActionProvider`