feat(testlab): add createRestAppClient(), simplify usage in tests
Clean up all test code (including examples in the documentation)
to use `Client` instead of `supertest.SuperTest<supertest.Test>`.

Remove `createClientForRestServer` because it was not used anywhere
and had the issue of not stopping the server after the test is done.

Add a new helper `createRestAppClient` instead, rework acceptance tests
(including the templates) to use this new helper.