feat: add "filter" parameter to "find" endpoints
Modify the CLI template and example controllers to accept a new filter
argument, leveraging the recently added `@param.query.object()` and
Filter schema.

Rename the controller-method argument accepting model data from
`${controllerName}` to `data` (e.g. `data` instead of `TodoController`).
I think this was an oversight in the implementation of the REST
Controller template.

Update the code snippets in documentation that are showing controller
classes to match the actual code used in the examples and produced
by our `lb4 controller` command.