feat(testlab): always use IPv4 host in HTTP server config
Modify `givenHttpServerConfig()` helper to always use `127.0.0.1`
as the host name where to listen.

Before this change, we were using IPv4 address on Travis CI only.
Unfortunately, the problem with malformed IPv6 networking setup is not
specific to Travis CI only, it's a problem in Docker itself.

With this change in place, applications tests should work out of the box
in any Docker-based environment.

This commit also modifies the way how custom configuration values are
handled by `givenHttpServerConfig`. Empty values for `host` or `port`
are ignored and replaced with helper defaults (host 127.0.0.1, port 0).