build: move travis tasks to .travis.yml to get timing info
Our CI build is becoming too long, it takes 7-10 minutes for a single
worker to finish. This commit is changing the configuration of Travis CI
in such way that individual build steps are defined in a way that allows
the CI server to understand them and measure how long they take.