chore: remove Node.js 10.0.0 from AppVeyor config
The version 10.0.0 is way too outdated, it makes little sense to test
it. Unfortunately we cannot upgrade to a newer 10.x version because
the CLI tests freeze on Windows :(

Also, Node.js 10.0.0 is shipping an outdated npm version that does not
support `npm ci` command that we are using to speed up install times.