fix(openapi-v3): preserve `additionalProperties: false`
By default, additional properties are allowed in all objects.

Before this change, our code converting JSON Schema to OpenAPI Schema
was discarding `additionalProperties` field, therefere making it
impossible to describe a schema that does not allow additional props.

This change is fixing the conversion code to preserve
`additionalProperties` value.

Additionaly, we also fix the code generating schema for Filter & Where
to set `additionalProperties` and to correctly apply default value
of model setting `strict` (which is `true`).

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>