chore: use http status code 307 instead of 304 for tests
Node 13.x seems to not generate response for redirect(304) and it results
in timeouts for tests.

Quoted from https://nodejs.org/api/http.html:
In the http module, the response body is omitted when the request is a
HEAD request. Similarly, the 204 and 304 responses must not include a
message body.