feat(cli): improve UX of multi-item selection
When asking the user to select 1 or more items, e.g. models to import,
require at least one item to be selected.

This prevents users from accidentally pressing ENTER instead of SPACE
to select an item from the list, in which case the prompt is resolved
with an empty list and thus the CLI command ends prematurely.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>