chore: add codeowners for tutorials
While reviewing a recent pull request, I noticed that it was assigned
to all team members because one of the tutorial doc pages was updated
as part of the changes. I find that suboptimal because then it's not
clear who is primarily responsible for reviewing the changes.

In this pull request, I am adding new CODEOWNERS entries for each
sub-directory in /docs/site/tutorials/