fix(repository): updateById behaves like updateAll when called with an undefined value for id
updateById behaves like updateAll when it is triggered with undefined value, and as a side effect instead of trying to update a single row/model it updates all the rows. undefined is not a valid value for id, hence the function should throw an appropriate error when undefined is passed in.