feat(context): pass resolution context to the factory function for toDynamicValue()
A provider class can use dependency injection to receive resolution-related
metadata such as `context` and `binding`. But the overhead to wrap a factory
function is not desired for some use cases. This PR introduces a lightweight
alternative using `toDynamicValue` as follows:

1. Add a parameter to the factory function to receive resolution context
2. Support a specialized class with a static `value` method that allows
parameter injection.

Some benchmark tests are added to `benchmark` to measure performances for
various flavors of the dynamic value resolution for bindings.