feat(eslint-config): upgrade to ESLint 7.x
BREAKING CHANGE: ESLint 7.x changes how plugins references in the shared
config are resolved. The following is quoted from
https://eslint.org/blog/2020/05/eslint-v7.0.0-released:

> ESLint will now resolve plugins relative to the entry configuration file.
This means that shared configuration files that are located outside the
project can now be colocated with the plugins they require.
> Starting in ESLint v7, configuration files and ignore files passed to
ESLint using the --config path/to/a-config and --ignore-path
path/to/a-ignore CLI flags, respectively, will resolve from the current
working directory rather than the file location. This allows for users to
utilize shared plugins without having to install them directly in their
project.