docs: move "Reserved binding keys" to "Reference guides"
Also remove "Overview" heading from the page. I feel the heading is
redundant when there are no other headings. Before this change,
the page was rendered with a single-item Table of Contents at the top,
which looked weird to me. After this change, the pages is rendered
as a single block with the page title as the only heading.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>