feat(metadata): improve handling of missing design-time type metadata
Improve the typings for `getDesignTypeForProperty` and
`getDesignTypeForMethod` to account for the fact that no metadata may
be available when the project is written in pure JavaScript or
not enabling TypeScript compiler option `emitDecoratorMetadata`.

BREAKING CHANGE: If you are consuming `@loopback/metadata` directly,
then you may need to update your code to handle the case when
design-time type metadata is not available. (The compiler will tell you
what places to fix.)

Regular LoopBack users should not be affected as long as they update
all LB packages together.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>