fix(cli): Change to lowercase the project name type by the user
For the project generator, when the user use a project name with uppercase letters, the project can not be create because NPM packages has a validation to no allow create packages with uppercase letters.
The prompt was update to add a function to transform the values typed by the user to lowecase version and use this value as the project name.