test(cli): forward generator errors back to Mocha tests
Replace all remaining places invoking generator from tests via
`helpers.run()` to use `executeGenerator` instead, to ensure that
generator errors are correctly propagated back via a rejected promise.

Before this change, a failed generator behaved as if it succeeded. A
test suite would continue and fail later when one of the assumptions
are not true. As a result, the test error messages were cryptic and
made troubleshooting difficult.

For example, when running the following command from `packages/cli`:

    CI=1 lb-mocha test/acceptance/app-run.acceptance.js < /dev/null

Before my change, the suite would fail as follows:

    1) app-generator (SLOW)
         "before all" hook: install dependencies for·
            "passes `npm test` for the generated project":
       ValidationError: No packages remain after filtering [ '@loopback/sandbox-app' ]
        at filterPackages (node_modules/@lerna/filter-packages/filter-packages.js:54:13)

With my change in place, the suite correctly fails in the scaffolding
step:

    1) app-generator (SLOW)
         "before all" hook: scaffold a new application for
            "passes `npm test` for the generated project":
       SyntaxError: Unexpected end of JSON input
        at JSON.parse (<anonymous>)
        at AppGenerator._readJSONFromStdin (lib/base-generator.js:132:19)

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>