fix(cli): handle empty JSON config
Before this change, when the config read from `stdin` was an empty
string, the generator would fail with the following message:

    SyntaxError: Unexpected end of JSON input
       at JSON.parse (<anonymous>)
       at AppGenerator._readJSONFromStdin (lib/base-generator.js:132:19)

With this fix in place, an empty JSON string is treated as an empty
config object `{}` and the generator can continue its work.

Please note this change is probably a partial fix only. AFAICT,
the application generator is still not correctly scaffolding the project
when running in a non-TTY mode, despite the fact that it reports success.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>