ci: run GitHub Actions only for `master` and pull requests
Improve GitHub Actions configuration to run CI only on pushes to master
or pull requests to master.

Specifically, the new configuration does not run CI for pushes to feature
branches, to avoid duplicate runs for the same commit: one run for the
pull request, another for the feature branch the pull request is from.

Signed-off-by: Miroslav Bajtoš <mbajtoss@gmail.com>