feat: allow one strategy to fail the authentication process
The current code allows a strategy to throw an error but it only fails
the authentication process if other ones either skip or fail. This
behavior does not allow a strategy to abort the process. The PR introduces
a flag `failOnError` for authentication component so that a stratgey can
abort the process by throwing an error.

Signed-off-by: Raymond Feng <enjoyjava@gmail.com>