fix(sequelize): defer to `toJSON` for hidden property exclusion
Previously, the package used a custom `excludeHiddenProps` function to exclude hidden fields.
But that prevented users from accessing those fields on backend too.
With this change, we now defer to the LoopBack Model's built-in toJSON method,
which already handles the exclusion of hidden properties when serializing data for a response.

Fixes: https://github.com/sourcefuse/loopback4-sequelize/issues/37

Signed-off-by: Shubham P <shubham.prajapat@sourcefuse.com>